requirements.txt
streamlit
llama-index
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-pinecone
pinecone-client
google-generativeai
python-dotenv
PyMuPDF
playwright
requests
